+++ src/main/java/absyn/Stm.java
@@ -4,6 +4,6 @@ import java.util.Map;
 
 public abstract class Stm extends AST {
 
-   public abstract void interp(Map<String, Integer> mem);
+   public abstract void interp(Map<String, Double> mem);
 
 }
